move getWordlist to helper
